Added scoring function for AI

+++ src/main/scala/ai/minimax/MiniMaxMessage.scala
@@ -4,4 +4,4 @@ import model.{MatchfieldModel, PlayerModel}
 
 trait MiniMaxMessage
 case class RequestMessage(columnNr:Option[Int], simulatedPlayer:PlayerModel, otherPlayer:PlayerModel, matchfield:MatchfieldModel[PlayerModel], depth:Int) extends MiniMaxMessage
-case class ResponseMessage(columnNr:Option[Int], score:Int) extends MiniMaxMessage
+case class ResponseMessage(columnNr:Option[Int], score:Double) extends MiniMaxMessage
